requirements.txt
Flask
gunicorn
werkzeug
psycopg2-binary==2.9.9